+++ Assets/Scripts/UI/Views/PlayView.cs
@@ -806,6 +806,7 @@ namespace Strategos.UI.Views
             var row2 = AddButtonRow(parent);
             AddButton(row2, "SKIRMISH ONLY", () => LoadScenario(ScenarioSamples.SkirmishName));
             AddButton(row2, "PUSH NORTH", () => LoadScenario(ScenarioSamples.PushNorthName));
+            AddButton(row2, "TUTORIAL", () => LoadScenario(ScenarioSamples.TutorialName));
 
             var row3 = AddButtonRow(parent);
             AddButton(row3, "SAVE", QuickSave);